pyproject.toml
[tool.poetry]
name = "news-project"
version = "0.1.0"
description = ""
authors = ["Nurlan-Aliev <nualiev612@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.12"}
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
